requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyjwt==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
